Move the header file out of source and into include folder inside rviz_plugin
